backend/requirements.txt
fastapi
fastapi[standard]
uvicorn[standard]
redis[asyncio]
asyncpg
sqlalchemy[asyncio]
alembic
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-dotenv
python-json-logger
